backend/requirements.txt
Flask==2.0.2
Flask-SQLAlchemy==2.5.1
SQLAlchemy==1.4.27
Flask-Migrate==3.1.0
Flask-Cors==3.0.10
qrcode==6.1
Pillow==8.4.0
Web3==5.23.1
matplotlib==3.4.3
